requirements.txt
# OBS Fog Server - Python Dependencies

# FastAPI framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0
python-dotenv==1.0.1

# HTTP client (for Telegram API)
httpx==0.28.1

# Rate limiting
slowapi==0.1.9

# Encryption (for Steam files)
cryptography==44.0.0

# Redis (for rate limiting & caching)
redis==5.2.1

# Templates (for web interface)
jinja2==3.1.5

# Utilities
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1
aiosqlite==0.20.0

# Development
black==24.10.0
isort==5.13.2
mypy==1.14.1
ruff==0.9.2
